+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 
  * Upgrade support to OpenCV 2.4.10
  * Upgrade supported FFmpeg API to the 2.5 release branch
+ * Fix `time_base` warnings displayed by `FFmpegFrameRecorder` ([issue #75](https://github.com/bytedeco/javacv/issues/75))
  * Add new `TemplateMatching` sample, thanks to Waldemar Neto
  * Update instructions in the `README.md` file for manual installation in Android Studio
  * Replace deprecated `CvMat` and `IplImage` functionality used in the `Demo` class with new `Indexer` API, and provide sample `pom.xml` file for Maven